pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astrology-tool"
version = "1.0.0"
description = "Astrology analysis tool combining Tử Vi Đẩu Số and Western Astrology with AI interpretation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Astrology Tool Team"}
]
keywords = ["astrology", "tuvi", "horoscope", "zodiac", "deepseek"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "pyswisseph>=2.10.0",
    "lunardate>=0.2.2",
    "geopy>=2.4.0",
    "timezonefinder>=6.2.0",
    "pytz>=2024.1",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "jinja2>=3.1.2",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]
pdf = [
    "reportlab>=4.0.0",
]

[project.scripts]
astrology = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
